+++ src/config/titulo.js
@@ -1,1 +1,2 @@
-module.exports = 'NOMBRE DEL COMPONENTE FORMATIVO'
+module.exports =
+  'Formulación de proyectos agropecuarios bajo el enfoque metodológico de marco lógico'